# Conflicts:
#   frontend/1s-before/package-lock.json
#   frontend/1s-before/package.json

+++ backend/src/main/java/com/ssafy/fiftyninesec/search/entity/Keyword.java
@@ -10,7 +10,7 @@ public class Keyword {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer keywordId;
+    private Long keywordId;
 
     @Column(nullable = false)
     private String word;
